requirements.txt
absl-py==0.4.1
astor==0.7.1
cairocffi==0.9.0
CairoSVG==2.1.3
cffi==1.11.5
Click==7.0
cssselect2==0.2.1
cycler==0.10.0
defusedxml==0.5.0
Django==2.0.8
Flask==1.1.1
gast==0.2.0
gcloud==0.17.0
googleapis-common-protos==1.5.3
grpcio==1.14.2
h5py==2.8.0
html5lib==1.0.1
httplib2==0.11.3
itsdangerous==1.1.0
Jinja2==2.11.1
jws==0.1.3
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.1.1
matplotlib==2.2.3
numpy==1.14.5
oauth2client==3.0.0
pdfrw==0.4
Pillow==5.2.0
pkg-resources==0.0.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
pycryptodome==3.4.3
pyparsing==2.2.0
Pyphen==0.9.5
Pyrebase==3.0.27
python-dateutil==2.7.3
python-jwt==2.0.1
pytz==2018.5
PyYAML==3.13
requests==2.11.1
requests-toolbelt==0.7.0
rsa==3.4.2
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
tinycss2==0.6.1
WeasyPrint==0.42.3
webencodings==0.5.1
Werkzeug==0.16.1
